サーピスパッケージのインタフェース定義がなくなるように修正
