+++ project-root/backend/src/main/java/com/readrecords/backend/service/EmailChangeService.java
@@ -1,5 +1,37 @@
 package com.readrecords.backend.service;
 
-public interface EmailChangeService {
-    void changeEmail(String userId, String newUserEmail);
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.readrecords.backend.repository.EmailChangeRepository;
+
+import jakarta.transaction.Transactional;
+
+@Service
+@Transactional
+public class EmailChangeService {
+    private static final Logger logger = LoggerFactory
+            .getLogger(EmailChangeService.class);
+
+    @Autowired
+    EmailChangeRepository emailChangeRepository;
+
+    // JavaMailSenderのインジェクション
+    @Autowired
+    private MailService mailService;
+
+    public void changeEmail(String userId, String newUserEmail) {
+        logger.info("Updating email for userId: {}", userId);
+        int updatedRows = emailChangeRepository.updateEmail(userId,
+                newUserEmail);
+        if (updatedRows == 0) {
+            throw new IllegalArgumentException(
+                    "User not found or email unchanged.");
+        }
+
+        // 共通のメール送信処理を呼び出す
+        mailService.sendMail(newUserEmail, "Libro Logからの確認メールです", "");
+    }
 }
